feat: 여행 장소(정보) 하트 수 조회

+++ src/main/java/com/meetravel/domain/place/entity/TravelPlaceEntity.java
@@ -45,7 +45,7 @@ public class TravelPlaceEntity extends BaseEntity {
 
     @Builder.Default
     @Column(name = "LIKE_CNT")
-    private int likeCount = 0;
+    private long likeCount = 0;
 
     public void incrementLikeCount() {
         this.likeCount++;
